In lieu of feature flags, use entities and triples route for showcasing new tabs and tables
